feat : member repository #20
